requirements.txt
instagrapi
woocommerce
openai
python-dotenv
fastapi
uvicorn
pydantic
pydantic-settings
requests
gunicorn
